web_framework/mydjango/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-redis==5.2.0
